refactor: refactoring get color palette
